+++ lib/fupeg.rb
@@ -2,10 +2,9 @@
 
 require_relative "fupeg/version"
 require_relative "fupeg/parser"
+require_relative "fupeg/grammar"
 
 module FuPeg
-  VERSION = "0.1.0"
-
   class Error < StandardError; end
   # Your code goes here...
 end